Desktop/requirements.txt
PyQt5 
PyQt5-tools
opencv-contrib-python
asyncio
firebase_admin
numpy
python-telegram-bot
